Use textEdit of completionItem unconditionally
The server's CompletionItem.textEdit should always be respected,
because in recent versions of the LSP, it is the server who should
tell what to replace and what not.

Before this change, a trigger character was always ignored, whether
the server requested to replace it or not. This could be problematic
in some cases.

For instance, in PHP, you have a dollar sign beforeevery variable.
If the server request to replace say `$th` with `$this` using a
TextEdit interface, the client translates this to replacing `th`
with `$this`, which results in a wrong `$$this` result.